Added separate pricing for contract vs donation
AL members have to contract the ISK to purchase ships,
  while ALI corp members can simply donate that isk to corp.

In the latter case the builder still needs to be paid by contracting
to corp, however.

ONce you're marked as a "corp member" the form will ask you if
you're in AL, or ALI via required radio button. This should be easy
for us to ensure they aren't just lying, as it'll tell an AL member
to donate isk when they clearly can't.

Also added order created date to orders#show, and a display
for the contract fee & total for AL members.

Fixes #17
Fixes #13